+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * As per previous point, dropping `xan scrape --sep`.
 * Dropping implicit unary function calls in moonblade pipelines. This feature was not well-known, confusing (an indentifier, could be understood as a call in a pipeline, only if not in first position...), and mostly useless now that moonblade has had a proper dot operator.
 * `xan plot -A/--aggregate` does not take an expression anymore but has an automatic selection of two modes: `sum` and `mean`. It should also be faster.
+* Renaming the `index` function as `row_index` for clarity.
 
 *Features*